--- laokou-common/laokou-common-context/src/main/java/org/laokou/common/context/util/User.java
@@ -1,97 +1,0 @@
-/*
- * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.laokou.common.context.util;
-
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import lombok.Builder;
-import org.jspecify.annotations.NullMarked;
-import org.jspecify.annotations.Nullable;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-
-import java.io.Serializable;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Set;
-
-/**
- * @param id 用户ID.
- * @param username 用户名.
- * @param password 密码.
- * @param avatar 头像.
- * @param superAdmin 超级管理员标识.
- * @param status 用户状态 0启用 1禁用.
- * @param mail 邮箱.
- * @param mobile 手机号.
- * @param tenantId 租户ID.
- * @param permissions 菜单权限标识集合.
- * @param deptId 部门ID.
- * @author laokou
- */
-@Builder
-@JsonTypeName("User")
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "@type")
-@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY, getterVisibility = JsonAutoDetect.Visibility.NONE,
-		isGetterVisibility = JsonAutoDetect.Visibility.NONE, creatorVisibility = JsonAutoDetect.Visibility.NONE)
-public record User(Long id, String username, String password, String avatar, Boolean superAdmin, Integer status,
-		String mail, String mobile, Long tenantId, Long deptId, Set<String> permissions, Set<Long> deptIds,
-		Long creator) implements Authentication, Serializable {
-
-	@Override
-	@NullMarked
-	public String getName() {
-		return this.username;
-	}
-
-	@Override
-	@NullMarked
-	public Collection<GrantedAuthority> getAuthorities() {
-		return AuthorityUtils.createAuthorityList(Collections.emptyList());
-	}
-
-	@Override
-	public @Nullable Object getCredentials() {
-		return this.username;
-	}
-
-	@Override
-	public @Nullable Object getDetails() {
-		return this.username;
-	}
-
-	@Override
-	public @Nullable Object getPrincipal() {
-		return this.username;
-	}
-
-	@Override
-	public boolean isAuthenticated() {
-		return true;
-	}
-
-	@Override
-	public void setAuthenticated(boolean isAuthenticated) throws IllegalArgumentException {
-		throw new UnsupportedOperationException("Cannot change authentication state");
-	}
-
-}
