+++ laokou-common/laokou-common-context/src/main/java/org/laokou/common/context/util/UserConvertor.java
@@ -27,23 +27,22 @@ public final class UserConvertor {
 	private UserConvertor() {
 	}
 
-	public static UserExtDetails toUserDetails(User user, Set<String> scopes) {
-		return DomainFactory.createUserDetails()
+	public static OAuth2AuthenticatedExtPrincipal toPrincipal(UserExtDetails userExtDetails, Set<String> scopes) {
+		return DomainFactory.createPrincipal()
 			.toBuilder()
-			.id(user.id())
-			.username(user.username())
-			.password(user.password())
-			.avatar(user.avatar())
-			.superAdmin(user.superAdmin())
-			.tenantId(user.tenantId())
-			.permissions(user.permissions())
+			.id(userExtDetails.id())
+			.username(userExtDetails.username())
+			.avatar(userExtDetails.avatar())
+			.superAdmin(userExtDetails.superAdmin())
+			.tenantId(userExtDetails.tenantId())
+			.permissions(userExtDetails.permissions())
 			.scopes(scopes)
-			.status(user.status())
-			.mail(user.mail())
-			.mobile(user.mobile())
-			.deptId(user.deptId())
-			.deptIds(user.deptIds())
-			.creator(user.creator())
+			.status(userExtDetails.status())
+			.mail(userExtDetails.mail())
+			.mobile(userExtDetails.mobile())
+			.deptId(userExtDetails.deptId())
+			.deptIds(userExtDetails.deptIds())
+			.creator(userExtDetails.creator())
 			.build()
 			.decryptUsername()
 			.decryptMail()
