+++ laokou-common/laokou-common-context/src/test/java/org/laokou/common/context/util/OAuth2AuthenticatedExtPrincipalTest.java
@@ -1,0 +1,317 @@
+/*
+ * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.laokou.common.context.util;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.laokou.common.crypto.util.AESUtils;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
+
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+
+/**
+ * UserExtDetails test class.
+ *
+ * @author laokou
+ */
+@DisplayName("UserExtDetails Unit Tests")
+class OAuth2AuthenticatedExtPrincipalTest {
+
+	private OAuth2AuthenticatedExtPrincipal principal;
+
+	@BeforeEach
+	void setUp() {
+		Set<String> permissions = Set.of("sys:user:query", "sys:user:add");
+
+		principal = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.username("testuser")
+			.scopes(Set.of("read", "write"))
+			.avatar("https://example.com/avatar.png")
+			.superAdmin(true)
+			.status(0)
+			.mail("test@example.com")
+			.mobile("13800138000")
+			.tenantId(100L)
+			.deptId(10L)
+			.permissions(permissions)
+			.build();
+	}
+
+	@Test
+	@DisplayName("Test builder creates UserExtDetails with all properties")
+	void test_builder_allProperties_createsUserExtDetails() {
+		// Then
+		Assertions.assertThat(principal.getId()).isEqualTo(1L);
+		Assertions.assertThat(principal.getUsername()).isEqualTo("testuser");
+		Assertions.assertThat(principal.getAvatar()).isEqualTo("https://example.com/avatar.png");
+		Assertions.assertThat(principal.getSuperAdmin()).isTrue();
+		Assertions.assertThat(principal.getStatus()).isZero();
+		Assertions.assertThat(principal.getMail()).isEqualTo("test@example.com");
+		Assertions.assertThat(principal.getMobile()).isEqualTo("13800138000");
+		Assertions.assertThat(principal.getTenantId()).isEqualTo(100L);
+		Assertions.assertThat(principal.getDeptId()).isEqualTo(10L);
+		Assertions.assertThat(principal.getPermissions())
+			.hasSize(2)
+			.containsExactlyInAnyOrder("sys:user:query", "sys:user:add");
+	}
+
+	@Test
+	@DisplayName("Test getAuthorities with valid permissions returns correct authorities")
+	void test_getAuthorities_withValidPermissions_returnsAuthorities() {
+		// When
+		Collection<? extends GrantedAuthority> authorities = principal.getAuthorities();
+
+		// Then
+		Assertions.assertThat(authorities)
+			.isNotNull()
+			.hasSize(4)
+			.extracting(GrantedAuthority::getAuthority)
+			.containsExactlyInAnyOrder("sys:user:query", "sys:user:add", "read", "write");
+	}
+
+	@Test
+	@DisplayName("Test getAttributes always returns empty map")
+	void test_getAttributes_always_returnsEmptyMap() {
+		// When
+		Map<String, Object> attributes = principal.getAttributes();
+
+		// Then
+		Assertions.assertThat(attributes).isNotNull().isEmpty();
+	}
+
+	@Test
+	@DisplayName("Test getName with valid username returns username")
+	void test_getName_withValidUsername_returnsUsername() {
+		// When
+		String name = principal.getName();
+
+		// Then
+		Assertions.assertThat(name).isEqualTo("testuser");
+	}
+
+	@Test
+	@DisplayName("Test toBuilder modifying properties creates new instance")
+	void test_toBuilder_modifyProperties_createsNewInstance() {
+		// When
+		OAuth2AuthenticatedExtPrincipal modified = principal.toBuilder().id(999L).username("newuser").build();
+
+		// Then
+		Assertions.assertThat(modified.getId()).isEqualTo(999L);
+		Assertions.assertThat(modified.getUsername()).isEqualTo("newuser");
+		// Original should be unchanged
+		Assertions.assertThat(principal.getId()).isEqualTo(1L);
+		Assertions.assertThat(principal.getUsername()).isEqualTo("testuser");
+	}
+
+	@Test
+	@DisplayName("Test class type is OAuth2AuthenticatedPrincipal implementation")
+	void test_classType_always_implementsUserDetails() {
+		// Then
+		Assertions.assertThat(principal).isInstanceOf(OAuth2AuthenticatedPrincipal.class);
+	}
+
+	@Test
+	@DisplayName("Test class type is OAuth2AuthenticatedPrincipal implementation")
+	void test_classType_always_implementsOAuth2AuthenticatedPrincipal() {
+		// Then
+		Assertions.assertThat(principal).isInstanceOf(OAuth2AuthenticatedPrincipal.class);
+	}
+
+	@Test
+	@DisplayName("Test equals with same key fields returns true")
+	void test_equals_sameKeyFields_returnsTrue() {
+		// Given
+		OAuth2AuthenticatedExtPrincipal sameUser = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.username("testuser")
+			.avatar("different_avatar")
+			.superAdmin(false)
+			.status(1)
+			.mail("different@example.com")
+			.mobile("different_mobile")
+			.tenantId(100L)
+			.deptId(10L)
+			.permissions(new HashSet<>())
+			.build();
+
+		OAuth2AuthenticatedExtPrincipal differentUser = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(2L)
+			.username("otheruser")
+			.tenantId(100L)
+			.deptId(10L)
+			.build();
+
+		// Then
+		Assertions.assertThat(principal).isEqualTo(sameUser);
+		Assertions.assertThat(principal.hashCode()).isEqualTo(sameUser.hashCode());
+		Assertions.assertThat(principal).isNotEqualTo(differentUser);
+	}
+
+	@Test
+	@DisplayName("Test builder with null values creates UserExtDetails with nulls")
+	void test_builder_nullValues_createsUserExtDetailsWithNulls() {
+		// Given
+		OAuth2AuthenticatedExtPrincipal nullUser = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(null)
+			.username(null)
+			.avatar(null)
+			.superAdmin(null)
+			.status(null)
+			.mail(null)
+			.mobile(null)
+			.tenantId(null)
+			.deptId(null)
+			.permissions(null)
+			.build();
+
+		// Then
+		Assertions.assertThat(nullUser.getId()).isNull();
+		Assertions.assertThat(nullUser.getUsername()).isNull();
+		Assertions.assertThat(nullUser.getAvatar()).isNull();
+		Assertions.assertThat(nullUser.getSuperAdmin()).isNull();
+		Assertions.assertThat(nullUser.getStatus()).isNull();
+		Assertions.assertThat(nullUser.getMail()).isNull();
+		Assertions.assertThat(nullUser.getMobile()).isNull();
+		Assertions.assertThat(nullUser.getTenantId()).isNull();
+		Assertions.assertThat(nullUser.getDeptId()).isNull();
+		Assertions.assertThat(nullUser.getPermissions()).isNull();
+	}
+
+	@Test
+	@DisplayName("Test decryptUsername with encrypted username returns decrypted")
+	void test_decryptUsername_encryptedUsername_returnsDecrypted() throws Exception {
+		// Given
+		String originalUsername = "admin";
+		String encryptedUsername = AESUtils.encrypt(originalUsername);
+
+		OAuth2AuthenticatedExtPrincipal userWithEncryptedUsername = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.username(encryptedUsername)
+			.build();
+
+		// When
+		OAuth2AuthenticatedExtPrincipal decrypted = userWithEncryptedUsername.decryptUsername();
+
+		// Then
+		Assertions.assertThat(decrypted.getUsername()).isEqualTo(originalUsername);
+	}
+
+	@Test
+	@DisplayName("Test decryptMail with encrypted mail returns decrypted")
+	void test_decryptMail_encryptedMail_returnsDecrypted() throws Exception {
+		// Given
+		String originalMail = "test@example.com";
+		String encryptedMail = AESUtils.encrypt(originalMail);
+
+		OAuth2AuthenticatedExtPrincipal userWithEncryptedMail = org.laokou.common.context.util.OAuth2AuthenticatedExtPrincipal
+			.builder()
+			.id(1L)
+			.mail(encryptedMail)
+			.build();
+
+		// When
+		OAuth2AuthenticatedExtPrincipal decrypted = userWithEncryptedMail.decryptMail();
+
+		// Then
+		Assertions.assertThat(decrypted.getMail()).isEqualTo(originalMail);
+	}
+
+	@Test
+	@DisplayName("Test decryptMobile with encrypted mobile returns decrypted")
+	void test_decryptMobile_encryptedMobile_returnsDecrypted() throws Exception {
+		// Given
+		String originalMobile = "13800138000";
+		String encryptedMobile = AESUtils.encrypt(originalMobile);
+
+		OAuth2AuthenticatedExtPrincipal userWithEncryptedMobile = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.mobile(encryptedMobile)
+			.build();
+
+		// When
+		OAuth2AuthenticatedExtPrincipal decrypted = userWithEncryptedMobile.decryptMobile();
+
+		// Then
+		Assertions.assertThat(decrypted.getMobile()).isEqualTo(originalMobile);
+	}
+
+	@Test
+	@DisplayName("Test decryptUsername with null username returns same instance")
+	void test_decryptUsername_nullUsername_returnsSameInstance() {
+		// Given
+		OAuth2AuthenticatedExtPrincipal userWithNullUsername = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.username(null)
+			.build();
+
+		// When
+		OAuth2AuthenticatedExtPrincipal result = userWithNullUsername.decryptUsername();
+
+		// Then
+		Assertions.assertThat(result).isSameAs(userWithNullUsername);
+		Assertions.assertThat(result.getUsername()).isNull();
+	}
+
+	@Test
+	@DisplayName("Test decryptUsername with empty username returns same instance")
+	void test_decryptUsername_emptyUsername_returnsSameInstance() {
+		// Given
+		OAuth2AuthenticatedExtPrincipal userWithEmptyUsername = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.username("")
+			.build();
+
+		// When
+		OAuth2AuthenticatedExtPrincipal result = userWithEmptyUsername.decryptUsername();
+
+		// Then
+		Assertions.assertThat(result).isSameAs(userWithEmptyUsername);
+	}
+
+	@Test
+	@DisplayName("Test chained decrypt methods returns fully decrypted instance")
+	void test_decryptChain_allFieldsEncrypted_returnsFullyDecrypted() throws Exception {
+		// Given
+		String originalUsername = "admin";
+		String originalMail = "admin@example.com";
+		String originalMobile = "13800138000";
+
+		OAuth2AuthenticatedExtPrincipal user = OAuth2AuthenticatedExtPrincipal.builder()
+			.id(1L)
+			.username(AESUtils.encrypt(originalUsername))
+			.mail(AESUtils.encrypt(originalMail))
+			.mobile(AESUtils.encrypt(originalMobile))
+			.build();
+
+		// When
+		OAuth2AuthenticatedExtPrincipal decrypted = user.decryptUsername().decryptMail().decryptMobile();
+
+		// Then
+		Assertions.assertThat(decrypted.getUsername()).isEqualTo(originalUsername);
+		Assertions.assertThat(decrypted.getMail()).isEqualTo(originalMail);
+		Assertions.assertThat(decrypted.getMobile()).isEqualTo(originalMobile);
+	}
+
+}
