+++ laokou-common/laokou-common-websocket/src/main/java/org/laokou/common/websocket/model/enums/WebSocketType.java
@@ -21,7 +21,7 @@ import io.netty.channel.Channel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.laokou.common.i18n.util.EnumParser;
-import org.laokou.common.context.util.UserExtDetails;
+import org.laokou.common.context.util.OAuth2AuthenticatedExtPrincipal;
 import org.laokou.common.websocket.config.WebSocketSessionHeartBeatManager;
 import org.laokou.common.websocket.config.WebSocketSessionManager;
 import org.laokou.common.websocket.model.WebSocketMessage;
@@ -35,16 +35,16 @@ public enum WebSocketType {
 
 	MESSAGE("message", "消息") {
 		@Override
-		public void handle(UserExtDetails userExtDetails, WebSocketMessage wsm, Channel channel) {
-			log.info("【WebSocket-Server】 => 接收到消息，通道ID：{}，用户ID：{}，消息：{}", channel.id().asLongText(),
-					userExtDetails.getId(), wsm.getPayload());
+		public void handle(OAuth2AuthenticatedExtPrincipal principal, WebSocketMessage wsm, Channel channel) {
+			log.info("【WebSocket-Server】 => 接收到消息，通道ID：{}，用户ID：{}，消息：{}", channel.id().asLongText(), principal.getId(),
+					wsm.getPayload());
 		}
 	},
 
 	CONNECT("connect", "建立连接") {
 		@Override
-		public void handle(UserExtDetails userExtDetails, WebSocketMessage wsm, Channel channel) {
-			Long clientId = userExtDetails.getId();
+		public void handle(OAuth2AuthenticatedExtPrincipal principal, WebSocketMessage wsm, Channel channel) {
+			Long clientId = principal.getId();
 			log.info("【WebSocket-Server】 => 已建立连接，通道ID：{}，用户ID：{}", channel.id().asLongText(), clientId);
 			WebSocketSessionManager.add(clientId, channel);
 		}
@@ -52,13 +52,13 @@ public enum WebSocketType {
 
 	PING("ping", "发送心跳") {
 		@Override
-		public void handle(UserExtDetails userExtDetails, WebSocketMessage wsm, Channel channel) {
+		public void handle(OAuth2AuthenticatedExtPrincipal principal, WebSocketMessage wsm, Channel channel) {
 		}
 	},
 
 	PONG("pong", "心跳应答") {
 		@Override
-		public void handle(UserExtDetails userExtDetails, WebSocketMessage wsm, Channel channel) {
+		public void handle(OAuth2AuthenticatedExtPrincipal principal, WebSocketMessage wsm, Channel channel) {
 			String channelId = channel.id().asLongText();
 			log.info("【WebSocket-Server】 => 接收{}心跳{}", channelId, wsm.getPayload());
 			if (WebSocketSessionHeartBeatManager.get(channelId) > 0) {
@@ -80,6 +80,6 @@ public enum WebSocketType {
 		return EnumParser.parse(WebSocketType.class, WebSocketType::getCode, code);
 	}
 
-	public abstract void handle(UserExtDetails userExtDetails, WebSocketMessage wsm, Channel channel);
+	public abstract void handle(OAuth2AuthenticatedExtPrincipal principal, WebSocketMessage wsm, Channel channel);
 
 }
