+++ laokou-service/laokou-auth/laokou-auth-infrastructure/src/main/java/org/laokou/auth/config/OAuth2AuthorizationServerConfig.java
@@ -249,4 +249,5 @@ class OAuth2AuthorizationServerConfig {
 		}
 		return code.equalsIgnoreCase(captcha.toString());
 	}
+
 }
