+++ laokou-service/laokou-auth/laokou-auth-infrastructure/src/main/java/org/laokou/auth/config/authentication/AbstractOAuth2AuthenticationProvider.java
@@ -24,13 +24,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.NonNull;
 import org.laokou.auth.model.AuthA;
 import org.laokou.auth.model.constant.OAuth2Constants;
+import org.laokou.common.context.util.UserExtDetails;
 import org.laokou.common.core.util.RequestUtils;
 import org.laokou.common.i18n.common.constant.StringConstants;
 import org.laokou.common.i18n.common.exception.GlobalException;
 import org.laokou.common.i18n.util.ObjectUtils;
 import org.laokou.common.security.handler.OAuth2ExceptionHandler;
 import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
 import org.springframework.security.oauth2.core.ClaimAccessor;
@@ -202,7 +202,7 @@ abstract class AbstractOAuth2AuthenticationProvider implements AuthenticationPro
 	 * @param authA 认证聚合根
 	 * @return 用户信息
 	 */
-	protected UsernamePasswordAuthenticationToken authentication(AuthA authA, HttpServletRequest request) {
+	protected UserExtDetails authentication(AuthA authA, HttpServletRequest request) {
 		return authenticationProcessor.authentication(authA, request);
 	}
